pyproject.toml
[project]
name = "sauce-labs-selenium-pytest"
version = "1.0.0"
description = "Proof of concept for Sauce Labs with Selenium and Pytest"
authors = [
    { name = "Mateo Agudelo Garcia", email = "mateo.agudelo80@eia.edu.co" },
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "selenium (>=4.29.0,<5.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-selenium (>=4.1.0,<5.0.0)",
    "allure-pytest (>=2.13.5,<3.0.0)",
    "allure-python-commons (>=2.13.5,<3.0.0)",
    "pytest-rerunfailures (>=15.0,<16.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "faker (>=37.0.0,<38.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "tests" }]

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"
pylint = "^3.3.5"
flake8 = "^7.1.2"

[tool.pytest.ini_options]
reruns = 3
testpaths = ["tests"]
addopts = ["--alluredir", "allure-results", "--clean-alluredir"]
selenium_capture_debug = "failure"
selenium_exclude_debug = "html:log:url"
